// emasaku-cloudflare--semua-menjadi-satu/types.ts
export enum UserRole {
  MEMBER = 'Member',
  ADMIN = 'Admin',
  DEVELOPER = 'Developer',
}

export interface User {
  id: string;
  name: string;
  email?: string;
  avatarUrl: string;
  role: UserRole;
  joinedDate: string;
  whatsapp?: string;
  driverBalance?: number;
  isRetailPartner?: boolean;
  isLaundryPartner?: boolean;
  isSeller?: boolean;
  isWastePartner?: boolean;
}